Add content of init file to new module
